# docs/swagger.yaml
basePath: /
definitions:
...
  api.contradictionsResponse:
    properties:
      contradictions:
        items:
          type: string
        type: array
      path:
        type: string
    type: object
...
  api.metaResponse:
    properties:
      count:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      results:
        items:
          $ref: '#/definitions/api.metaResultEntry'
        type: array
    type: object
  api.metaResultEntry:
    properties:
      frontmatter:
        additionalProperties: {}
        type: object
      path:
        type: string
      permalink:
        type: string
    type: object
...
  api.searchResponse:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      query:
        type: string
      results:
        items:
          $ref: '#/definitions/api.searchResultEntry'
        type: array
      suggestions:
        items:
          $ref: '#/definitions/api.searchSuggestionEntry'
        type: array
    type: object
  api.searchResultEntry:
    properties:
      matches:
        items:
          $ref: '#/definitions/search.Match'
        type: array
      path:
        type: string
      permalink:
        type: string
      score:
        type: number
      snippet:
        type: string
    type: object
  api.searchSuggestionEntry:
    properties:
      distance:
        type: integer
      path:
        type: string
      query:
        type: string
      title:
        type: string
    type: object
...
  api.semanticRequest:
    properties:
      modifiedAfter:
        type: string
      offset:
        type: integer
      query:
        type: string
      topK:
        type: integer
    type: object
  api.semanticResponse:
    properties:
      offset:
        type: integer
      query:
        type: string
      results:
        items:
          $ref: '#/definitions/vectorstore.Result'
        type: array
      topK:
        type: integer
    type: object
...
  api.staleResponse:
    properties:
      count:
        type: integer
      results:
        items:
          $ref: '#/definitions/search.MetaResult'
        type: array
      staleDays:
        type: integer
    type: object
...
  search.Match:
    properties:
      line:
        type: integer
      text:
        type: string
    type: object
  search.MetaResult:
    properties:
      frontmatter:
        additionalProperties: {}
        type: object
      path:
        type: string
    type: object
...
  vectorstore.Result:
    properties:
      chunkIdx:
        type: integer
      path:
        type: string
      score:
        type: number
      snippet:
        type: string
    type: object
...
host: localhost:3333
info:
  contact:
    name: KiwiFS Support
    url: https://github.com/kiwifs/kiwifs
  description: KiwiFS is a content-addressable file system with an API.
  license:
    name: Business Source License 1.1
    url: https://github.com/kiwifs/kiwifs/blob/main/LICENSE
  title: KiwiFS API
  version: "1.0"
paths:
...
  /api/kiwi/contradictions:
    get:
      description: Finds pages that contradict the content of the specified page.
      parameters:
      - description: Path of the file to check for contradictions (must start with
          '/')
        in: query
        name: path
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.contradictionsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Find page contradictions
      tags:
      - search
...
  /api/kiwi/meta:
    get:
      description: Queries page metadata using filters, sorting, and pagination.
      parameters:
      - collectionFormat: csv
        description: Repeatable JSON-path filters (e.g. '$.tags[*]=project')
        in: query
        items:
          type: string
        name: where
        type: array
      - collectionFormat: csv
        description: OR filters in the format 'field operator value'
        in: query
        items:
          type: string
        name: or
        type: array
      - description: Field to sort the results by
        in: query
        name: sort
        type: string
      - description: Sorting order ('asc' or 'desc')
        in: query
        name: order
        type: string
      - description: Maximum number of results to return
        in: query
        name: limit
        type: integer
      - description: Number of results to skip (offset)
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.metaResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Query page metadata
      tags:
      - search
...
  /api/kiwi/search:
    get:
      description: Performs a full-text search query on the repository content.
      parameters:
      - description: Search query string
        in: query
        name: q
        required: true
        type: string
      - description: 'Maximum number of search results to return (default: 15, max:
          200)'
        in: query
        name: limit
        type: integer
      - description: 'Number of search results to skip (offset) (default: 0)'
        in: query
        name: offset
        type: integer
      - description: Set to 'none' or 'off' to disable trust boosting in search results
        in: query
        name: boost
        type: string
      - description: RFC3339 formatted cutoff date to filter search results by modification
          time
        in: query
        name: modifiedAfter
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.searchResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Full-text search
      tags:
      - search
  /api/kiwi/search/semantic:
    post:
      consumes:
      - application/json
      description: Performs a semantic/vector search using embeddings. Accepts query
        parameters or a JSON body.
      parameters:
      - description: JSON body containing query, topK, offset, and modifiedAfter filters
        in: body
        name: body
        schema:
          $ref: '#/definitions/api.semanticRequest'
      - description: Search query string (used if not provided in JSON body)
        in: query
        name: q
        type: string
      - description: 'Maximum number of search results to return (default: 10)'
        in: query
        name: topK
        type: integer
      - description: Number of search results to skip (offset)
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.semanticResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Semantic vector search
      tags:
      - search
  /api/kiwi/search/verified:
    get:
      description: Performs a verified full-text search query on the repository using
        the SQLite search backend.
      parameters:
      - description: Search query string
        in: query
        name: q
        required: true
        type: string
      - description: 'Maximum number of search results to return (default: 15, max:
          200)'
        in: query
        name: limit
        type: integer
      - description: 'Number of search results to skip (offset) (default: 0)'
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.searchResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Verified full-text search
      tags:
      - search
...
  /api/kiwi/stale:
    get:
      description: Returns pages that have not been modified for a given number of
        days.
      parameters:
      - description: Number of days threshold for staleness (default 30)
        in: query
        name: days
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.staleResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "501":
          description: Not Implemented
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get stale pages
      tags:
      - search
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your API key.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"